Moved some functions from fragments to viewmodel
